Clean up and Refactored
